adding optimistic locking feature

+++ src/main/java/com/upgrade/controller/ReservationController.java
@@ -1,10 +1,10 @@
 package com.upgrade.controller;
 
 import java.net.URI;
-
 import java.text.ParseException;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -18,18 +18,19 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.util.UriComponentsBuilder;
 
+import com.upgrade.config.Constants;
 import com.upgrade.domain.Reservation;
 import com.upgrade.dto.ReservationDto;
+import com.upgrade.repository.ReservationDao;
 import com.upgrade.service.ReservationService;
 import com.upgrade.service.validation.ValidationException;
 import com.upgrade.utils.NotFoundException;
 import com.upgrade.utils.Utils;
-import com.upgrade.config.Constants;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -46,6 +47,9 @@ public class ReservationController {
 
 	@Autowired
 	ReservationService reservationService;
+
+	@Autowired
+	ReservationDao reservationDao;
 
 	@GetMapping(Constants.AVAILABILITIES)
 	@ApiOperation(value = "Retrieve the availabilities for a future reservation")
@@ -131,17 +135,19 @@ public class ReservationController {
 		URI locationUri = ucb.path(Constants.ROOT_API).path("/").path(String.valueOf(dto.getIdentifier())).build()
 				.toUri();
 		hdrs.setLocation(locationUri);
-		ResponseEntity<ReservationDto> responseEntity = new ResponseEntity<ReservationDto>(dto, hdrs, HttpStatus.OK);
+		hdrs.setETag(Long.toString(reservation.getVersion()));
 		
+		ResponseEntity<ReservationDto> responseEntity = new ResponseEntity<ReservationDto>(dto, hdrs, HttpStatus.OK);
+
 		return responseEntity;
 	}
 
 	@RequestMapping(path = "", method = RequestMethod.POST, consumes = "application/json;charset=UTF-8", produces = "application/json;charset=UTF-8")
 	@ApiOperation(value = "Add a reservation for a given customer after validation")
-	public ResponseEntity<Reservation> addBooking(@RequestBody ReservationDto res, UriComponentsBuilder ucb)
+	public synchronized ResponseEntity<Reservation> addBooking(@RequestBody ReservationDto res, UriComponentsBuilder ucb)
 			throws ParseException, ValidationException {
+		
 		Reservation reserv = null;
-
 
 		Validate.notNull(res, "reservation object must not be null");
 		
@@ -154,9 +160,8 @@ public class ReservationController {
 			throw new ValidationException(e.getMessage());
 		}
 
-		HttpHeaders hdrs = new HttpHeaders();
-		URI locationUri = ucb.path(Constants.ROOT_API).build().toUri();
-		hdrs.setLocation(locationUri);
+		HttpHeaders hdrs = createHeaders(ucb, reserv);
+		
 		ResponseEntity<Reservation> responseEntity = new ResponseEntity<Reservation>(reserv, hdrs, HttpStatus.CREATED);
 		
 		return responseEntity;
@@ -164,15 +169,29 @@ public class ReservationController {
 
 	@PutMapping("/{identifier}")
 	@ApiOperation(value = "Modify a reservation for a given customer after validation")
-	public ResponseEntity<Reservation> modifyBooking(@PathVariable("identifier") String identifier, @RequestBody ReservationDto res,
+	public synchronized ResponseEntity<Reservation> modifyBooking(WebRequest request, @PathVariable("identifier") String identifier, @RequestBody ReservationDto res,
 			UriComponentsBuilder ucb) throws ParseException, ValidationException {
-
-		Reservation reserv = null;
-
 		Validate.notNull(res, "reservation object must not be null");
 		Validate.notNull(identifier, "identifier object must not be null");
 
 		log.info("ModifyBooking called for identifier [{}}", identifier);
+		
+		Reservation reserv =  reservationService.getReservationByIdentifier(identifier);
+		if (null == reserv) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		res.setIdentifier(identifier);
+		
+		String ifMatchValue = request.getHeader("If-Match");
+		
+		if (StringUtils.isEmpty(ifMatchValue)) {
+			return ResponseEntity.badRequest().build();
+		}
+		
+		if (!Long.valueOf(ifMatchValue.replace("\"", "")).equals(reserv.getVersion()) ){
+			return ResponseEntity.status(HttpStatus.PRECONDITION_FAILED).build();
+		}
 		
 		try {
 			reserv = reservationService.modifyReservation(res, identifier);
@@ -182,41 +201,61 @@ public class ReservationController {
 			throw new ValidationException(e.getMessage());
 		}
 
-		HttpHeaders hdrs = new HttpHeaders();
+		HttpHeaders hdrs = createHeaders(ucb, reserv);
 		
-		URI locationUri = ucb.path(Constants.ROOT_API).path("/").path(String.valueOf(reserv.getIdentifier())).build()
-				.toUri();
-		hdrs.setLocation(locationUri);
-		ResponseEntity<Reservation> responseEntity = new ResponseEntity<Reservation>(reserv, hdrs, HttpStatus.CREATED);
+		ResponseEntity<Reservation> responseEntity = new ResponseEntity<Reservation>(reserv, hdrs, HttpStatus.OK);
 		
 		return responseEntity;
 	}
 
 	@DeleteMapping("/{identifier}")
 	@ApiOperation(value = "Delete a reservation for a given identifier")
-	public ResponseEntity<ReservationDto> deleteBooking(@PathVariable("identifier") String identifier,
+	public synchronized ResponseEntity<ReservationDto> deleteBooking( WebRequest request, @PathVariable("identifier") String identifier,
 			@RequestBody ReservationDto res, UriComponentsBuilder ucb) throws ParseException, ValidationException {
 
 		Validate.notNull(res, "reservation object must not be null");
 		Validate.notNull(identifier, "identifier object must not be null");
 		log.info("DeleteBooking called for identifier [{}}", identifier);
-
+		
+		Reservation reserv =  reservationService.getReservationByIdentifier(identifier);
+		
+		if (null == reserv) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		String ifMatchValue = request.getHeader("If-Match");
+		
+		if (StringUtils.isEmpty(ifMatchValue)) {
+			return ResponseEntity.badRequest().build();
+		}
+		
+		if (!Long.valueOf(ifMatchValue.replace("\"", "")).equals(reserv.getVersion()) ){
+			return ResponseEntity.status(HttpStatus.PRECONDITION_FAILED).build();
+		}
+		
 		try {
 			reservationService.deleteReservation(res, identifier);
 
 		} catch (NotFoundException e) {
-			log.info("ValidationException when retrieving availabilities:[{}]", e.getMessage());
+			log.info("ValidationException when deleting a reservation:[{}]", e.getMessage());
 			throw new ValidationException(e.getMessage());
 		}
 
-		HttpHeaders hdrs = new HttpHeaders();
-		URI locationUri = ucb.path(Constants.ROOT_API).path("/").path(String.valueOf(identifier)).build().toUri();
-		hdrs.setLocation(locationUri);
+		HttpHeaders hdrs = createHeaders(ucb, reserv);
 
 		ResponseEntity<ReservationDto> responseEntity = new ResponseEntity<ReservationDto>(null, hdrs,
-				HttpStatus.ACCEPTED);
+				HttpStatus.OK);
 		
 		return responseEntity;
+	}
+	
+	private HttpHeaders createHeaders(UriComponentsBuilder ucb, Reservation reserv) {
+
+		HttpHeaders hdrs = new HttpHeaders();
+		URI locationUri = ucb.path(Constants.ROOT_API).path("/").path(String.valueOf(reserv.getIdentifier())).build().toUri();
+		hdrs.setLocation(locationUri);
+		hdrs.setETag("\""+Long.toString(reserv.getVersion())+"\"");
+		return hdrs;
 	}
 
 	@ExceptionHandler(NotFoundException.class)
